don't use arena allocation
This doesn't actually seem to make memory management
any easier. Using std::unique_ptr seems to be a better
approach.

Note that due to the way bison works, it's probably not
going to be straightforward to avoid leaking the partial
parse tree or AST if there is a syntax error.